requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
tqdm>=4.62.0
Pillow>=9.0.0

# Optional: CLIP for embedding generation
# Uncomment one of the following:
# open_clip_torch>=2.20.0
# git+https://github.com/openai/CLIP.git

# Optional: Hyperparameter tuning
# optuna>=3.0.0

# Development
# pytest>=7.0.0
# black>=22.0.0
# isort>=5.10.0
